+++ api/v1/src/controllers/users.controller.ts
@@ -3,6 +3,9 @@ import User from '../models/User';
 
 export const getUsers: RequestHandler = async (req, res) => {
   try {
+    if (!req.registered)
+      return res.status(403).json({ message: "No token provided" });
+    
     const users = await User.find();
     res.json(users);
   } catch(error) {
@@ -11,12 +14,16 @@ export const getUsers: RequestHandler = async (req, res) => {
 };
 
 export const getUser: RequestHandler = async (req, res) => {
+  if (!req.registered)
+      return res.status(403).json({ message: "No token provided" });
   const user = await User.findById(req.params.id);
   if (!user) return res.status(204).json(); // no sale mensaje para el error 204
   res.json(user);
 };
 
 export const deleteUser: RequestHandler = async (req, res) => {
+  if (!req.registered)
+      return res.status(403).json({ message: "No token provided" });
   const user = await User.findByIdAndDelete(req.params.id);
   if (!user) return res.status(204).json(); // no sale mensaje para el error 204
   res.json(user);
@@ -24,6 +31,8 @@ export const deleteUser: RequestHandler = async (req, res) => {
 
 export const updateUser: RequestHandler = async (req: Request, res: Response) => {
   try {
+    if (!req.registered)
+      return res.status(403).json({ message: "No token provided" });
     const userFound = await User.findById(req.userId);
     if (!userFound) {
       return res.status(301).json({message: 'invalid user ID'});
@@ -45,8 +54,6 @@ export const updateUser: RequestHandler = async (req: Request, res: Response) =>
     {
       return res.status(301).json({message: 'email or username already exists'});
     }
-    console.log(req.userId);
-    console.log(req.body);
 
     const user = await User.findByIdAndUpdate(req.userId, req.body, { new: true });
     if (!user) return res.status(204).json(); // no sale mensaje para el error 204
@@ -57,8 +64,9 @@ export const updateUser: RequestHandler = async (req: Request, res: Response) =>
 };
 
 export const getProfile: RequestHandler = async (req, res) => {
+  if (!req.registered)
+      return res.status(403).json({ message: "No token provided" });
   const user = await User.findById(req.userId, {password: 0});
   if (!user) return res.status(204).json(); // no sale mensaje para el error 204
   res.json(user);
 };
-
